warnings fixed

+++ clover/src/PID.h
@@ -10,23 +10,23 @@
 class PID
 {
 public:
-    PID(double kp, double ki, double kd,
-        double min_out = -1e6, double max_out = 1e6,
-        double min_integral = -1e6, double max_integral = 1e6,
-        double integral_zone = std::numeric_limits<double>::infinity(),
-        double deriv_lowpass_hz = 0.0)
+    PID(float kp, float ki, float kd,
+        float min_out = -1e6f, float max_out = 1e6f,
+        float min_integral = -1e6f, float max_integral = 1e6f,
+        float integral_zone = std::numeric_limits<float>::infinity(),
+        float deriv_lowpass_hz = 0.0f)
         : kp_(kp), ki_(ki), kd_(kd)
     {
-        if (min_out != -1e6 || max_out != 1e6) {
+        if (min_out != -1e6f || max_out != 1e6f) {
             setOutputLimits(min_out, max_out);
         }
-        if (min_integral != -1e6 || max_integral != 1e6) {
+        if (min_integral != -1e6f || max_integral != 1e6f) {
             setIntegralLimits(min_integral, max_integral);
         }
-        if (integral_zone != std::numeric_limits<double>::infinity()) {
+        if (integral_zone != std::numeric_limits<float>::infinity()) {
             setIntegralZone(integral_zone);
         }
-        if (deriv_lowpass_hz > 0.0) {
+        if (deriv_lowpass_hz > 0.0f) {
             setDerivativeLowPass(deriv_lowpass_hz);
         }
     }
@@ -38,7 +38,7 @@ public:
     // dt is computed internally from steady_clock;
     // on the first call, derivative and integral are not applied to avoid a large transient.
 
-    double calculate(double setpoint, double measurement, double measurement_d, double dt)
+    float calculate(float setpoint, float measurement, float measurement_d, float dt)
     {
 
 
@@ -47,11 +47,11 @@ public:
 
             prev_meas_ = measurement;
             // First call: just proportional action.
-            const double error = setpoint - measurement;
+            const float error = setpoint - measurement;
             return clampOutput(kp_ * error);
         }
 
-        const double error = setpoint - measurement;
+        const float error = setpoint - measurement;
 
         // Integral (respect integral zone if enabled)
         if (!use_integral_zone_ || std::abs(error) <= integral_zone_)
@@ -60,43 +60,43 @@ public:
         }
 
         // Derivative on measurement to reduce derivative kick
-        double deriv_raw = 0.0;
+        float deriv_raw = 0.0f;
         if (!std::isnan(prev_meas_))
         {
-            deriv_raw = -measurement_d / std::max(dt, 1e-9); // negative sign: d(error)/dt = -d(meas)/dt
+            deriv_raw = -measurement_d / std::max(dt, 1e-9f); // negative sign: d(error)/dt = -d(meas)/dt
         }
         prev_meas_ = measurement;
 
         // Optional 1st-order low-pass on derivative: y += a*(x - y)
-        double deriv_term = deriv_raw;
+        float deriv_term = deriv_raw;
         if (use_deriv_lp_)
         {
-            const double rc = 1.0 / (2.0 * pi * std::max(deriv_cutoff_hz_, 1e-6));
-            const double a = dt / (rc + dt);
+            const float rc = 1.0f / (2.0f * pi * std::max(deriv_cutoff_hz_, 1e-6f));
+            const float a = dt / (rc + dt);
             deriv_state_ += a * (deriv_raw - deriv_state_);
             deriv_term = deriv_state_;
         }
 
-        double integral_output = integral_ * ki_;
+        float integral_output = integral_ * ki_;
         if (use_integral_limits_)
         {
             integral_output = std::clamp(integral_output, min_integral_, max_integral_);
         }
         // prevent division by 0
-        if (ki_ > 0.000001)
+        if (ki_ > 0.000001f)
         {
             integral_ = integral_output / ki_; // store back clamped integral
         }
-        double output = kp_ * error + integral_output + kd_ * deriv_term;
+        float output = kp_ * error + integral_output + kd_ * deriv_term;
 
         return clampOutput(output);
     }
 
     // uses previous measurement to calculate derivative
-    double calculate(double setpoint, double measurement, double dt)
+    float calculate(float setpoint, float measurement, float dt)
     {
         // Derivative on measurement to reduce derivative kick
-        double dmeas = 0.0;
+        float dmeas = 0.0f;
         if (!std::isnan(prev_meas_))
         {
             dmeas = measurement - prev_meas_;
@@ -107,15 +107,15 @@ public:
     }
 
     // Reset internal state (integral, derivative filter). Optionally set a new integral value.
-    void reset(double integral = 0.0)
+    void reset(float integral = 0.0f)
     {
         integral_ = integral;
-        prev_meas_ = std::numeric_limits<double>::quiet_NaN();
-        deriv_state_ = 0.0;
+        prev_meas_ = std::numeric_limits<float>::quiet_NaN();
+        deriv_state_ = 0.0f;
     }
 
     // --- Configuration helpers ---
-    void setGains(double kp, double ki, double kd)
+    void setGains(float kp, float ki, float kd)
     {
         kp_ = kp;
         ki_ = ki;
@@ -123,16 +123,16 @@ public:
         reset();
     }
 
-    void setP(double kp) { kp_ = kp; }
-    void setI(double ki) { ki_ = ki; }
-    void setD(double kd) { kd_ = kd; }
+    void setP(float kp) { kp_ = kp; }
+    void setI(float ki) { ki_ = ki; }
+    void setD(float kd) { kd_ = kd; }
 
-    double getP() const { return kp_; }
-    double getI() const { return ki_; }
-    double getD() const { return kd_; }
-    double getIntegralOutput() const { return integral_ * ki_; }
+    float getP() const { return kp_; }
+    float getI() const { return ki_; }
+    float getD() const { return kd_; }
+    float getIntegralOutput() const { return integral_ * ki_; }
 
-    void setOutputLimits(double min_out, double max_out)
+    void setOutputLimits(float min_out, float max_out)
     {
         min_out_ = min_out;
         max_out_ = max_out;
@@ -142,7 +142,7 @@ public:
     }
     void clearOutputLimits() { use_output_limits_ = false; }
 
-    void setIntegralLimits(double min_i, double max_i)
+    void setIntegralLimits(float min_i, float max_i)
     {
         min_integral_ = min_i;
         max_integral_ = max_i;
@@ -152,15 +152,15 @@ public:
     }
     void clearIntegralLimits() { use_integral_limits_ = false; }
 
-    void setIntegralZone(double zone_abs_error)
+    void setIntegralZone(float zone_abs_error)
     {
-        integral_zone_ = std::max(0.0, zone_abs_error);
+        integral_zone_ = std::max(0.0f, zone_abs_error);
         use_integral_zone_ = true;
     }
     void clearIntegralZone() { use_integral_zone_ = false; }
 
     // Enable a lowâ€‘pass filter on the derivative term. Example: cutoff_hz = 30.0
-    void setDerivativeLowPass(double cutoff_hz)
+    void setDerivativeLowPass(float cutoff_hz)
     {
         deriv_cutoff_hz_ = cutoff_hz;
         use_deriv_lp_ = true;
@@ -168,11 +168,11 @@ public:
     void clearDerivativeLowPass()
     {
         use_deriv_lp_ = false;
-        deriv_state_ = 0.0;
+        deriv_state_ = 0.0f;
     }
 
 private:
-    double clampOutput(double u) const
+    float clampOutput(float u) const
     {
         if (!use_output_limits_)
             return u;
@@ -184,27 +184,27 @@ private:
     }
 
     // Gains
-    double kp_{};
-    double ki_{};
-    double kd_{};
+    float kp_{};
+    float ki_{};
+    float kd_{};
 
     // State
-    const double pi = std::acos(-1.0);
+    const float pi = std::acos(-1.0);
 
-    double integral_ = 0.0;
-    double prev_meas_ = std::numeric_limits<double>::quiet_NaN();
+    float integral_ = 0.0f;
+    float prev_meas_ = std::numeric_limits<float>::quiet_NaN();
 
     // Derivative lowâ€‘pass state
     bool use_deriv_lp_ = false;
-    double deriv_cutoff_hz_ = 0.0;
-    double deriv_state_ = 0.0;
+    float deriv_cutoff_hz_ = 0.0f;
+    float deriv_state_ = 0.0f;
 
 
     // Limits
     bool use_output_limits_ = false;
-    double min_out_ = -1.0, max_out_ = 1.0;
+    float min_out_ = -1.0f, max_out_ = 1.0f;
     bool use_integral_limits_ = false;
-    double min_integral_ = -1e6, max_integral_ = 1e6;
+    float min_integral_ = -1e6f, max_integral_ = 1e6f;
     bool use_integral_zone_ = false;
-    double integral_zone_ = std::numeric_limits<double>::infinity();
+    float integral_zone_ = std::numeric_limits<float>::infinity();
 };
